integration: Move all container-related stuff into a pkg
